crawler改进1,使用houzz配合测试;baokuanyi简单代码提交

+++ crawler2.py
@@ -8,21 +8,39 @@ import socket
 import lockfile
 import os
 import shutil
+import time
+import random
+import ConfigParser
 import logging
 from logging import debug, info, warning, error
 from utils import read_config, print_csvline, load_http_headers
 
 
 class Merchant:
-    def __init__(self, merchantName, url):
+    def __init__(self, merchantName, url, crawlEntryUrl=None):
         self.merchantName = merchantName
         self.wwwDomain = url
         self.homeUrl = "http://" + url
+        self.headers = {}
+        self.crawlEntryUrl = crawlEntryUrl if crawlEntryUrl else self.homeUrl
+        self.sleepSeconds = 2, 5
         self.productFields = ['level1_category', 'level2_category', 'level3_category',\
                                'name', 'product_url', 'img_url',\
                                'sku_id', 'price', 'reviews']
-        self.crawlEntryUrl = None
+        self.maxPagesPerCategory = 999
+        self.useDistributedFetch = False
+        self.useWebClient = False
+        self.isSavePages = False
         self.config = None
+    
+    def initStoreDir(self):
+        rel_path = self.config.get2('savePagesDir')
+        if not rel_path:
+            error("No store dir configured to store pages... EXIT!")
+            raise Exception("No store dir configured to store pages... EXIT!")
+        abs_path = os.path.join(os.getcwd(), os.path.join(rel_path, self.merchantName))
+        if not os.path.exists(abs_path):
+            os.makedirs(abs_path)
     
     def filteruri(self, uri):
         if uri.startswith("http"):
@@ -56,6 +74,18 @@ class CategoryInfo:
     def getLevel3Category(self):
         if self.parent_categories and len(self.parent_categories) == 2:
             return self.name
+    
+    def formalize(self):
+        self.url = self.merchant.filteruri(self.url)
+        return self
+    
+    def __str__(self):
+        cate_str = self.getLevel1Category()
+        if self.getLevel2Category():
+            cate_str += (' > ' + self.getLevel2Category())
+            if self.getLevel3Category():
+                cate_str += (' > ' + self.getLevel3Category())
+        return cate_str
   
 class ProductInfo():
     def __init__(self, merchant):
@@ -80,8 +110,17 @@ class ProductInfo():
     def to_list(self):
         result = []
         for field in self.merchant.productFields:
-            result.append(self[field])
+            result.append(self.item_encode(self[field]))
         return result
+    
+    def item_encode(self, item):
+        if not item: #''
+            return ''
+        elif isinstance(item, unicode):
+            return item.encode('utf-8')
+        else:
+            print type(item)
+            return str(item)
 
 class FetchTask:
     def __init__(self, **kwargs):
@@ -98,6 +137,9 @@ class FetchTask:
 class Fetcher:
     def __init__(self, merchant):
         self.merchant = merchant
+    
+    def sleep(self):
+        time.sleep(random.uniform(2, 5))
         
     def fetch(self, task):
         h = httplib2.Http(timeout=30)
@@ -110,6 +152,7 @@ class Fetcher:
         status, content = 200, ''
         times = 0
         while True:
+            self.sleep()
             times += 1
             try:
                 status, content = self.fetch(fetchTask)
@@ -124,86 +167,145 @@ class Fetcher:
             if times >= try_times:
                 break
         return status, content
+
+class RabbitmqFetcher(Fetcher):
+    pass
+
+class Crawler:
+    def __init__(self, merchant, parser):
+        self.merchant = merchant
+        self.parser = parser
+        self.config = read_config('config/crawl.cfg')
+        if self.merchant.useDistributedFetch and self.config.get2('rabbitmq.server', section=''):
+            self.fetcher = RabbitmqFetcher()
+        else:
+            self.fetcher = Fetcher(merchant)
+        self.filename = merchant.merchantName + '_' + datetime.date.today().strftime("%m-%d-%Y") + '_productInfo.inprogress'
+        self.fp = None
     
+    def rmCleanFile(self):
+        pass
+
+    def initDataFile(self):
+        save_dir = self.config.get('all', 'save.dir')
+        self.filePath = os.path.join(save_dir, self.filename)
+        self.fp = open(self.filePath, 'w')
+
+    def crawl(self):
+        #删除空文件
+        self.rmCleanFile()
+        #初始化文件, 复用已经抓取的数据
+        self.initDataFile()
+        #从入口页获取可以获得的全部品类
+        categoryList = self.getInitialCategories()
+        #根据url进行一次去重
+        categoryList = self.rmDuplicateCategories(categoryList)
+        info("From the entrance page, we get following categories, totally %s:" %len(categoryList))
+        for category in categoryList:
+            info(category)
+        try:
+            self.crawlCategoryList(categoryList)
+        except Exception, e:
+            print e
+        finally:
+            self.fp.close()
+        
+        info('crawl end.')
+
+
+    def getInitialCategories(self):
+        homepage_content =self.fetchCategoryPageContent(self.merchant.crawlEntryUrl)
+        categoryList = self.parser.parseCategories(homepage_content)
+        return categoryList
+    
+    def rmDuplicateCategories(self, categoryList):
+        newList = []
+        urls = {}
+        for category in categoryList:
+            if not urls.has_key(category.url):
+                newList.append(category)
+                urls[category.url] = category
+        return newList
+    
+    def crawlCategoryList(self, categoryList):
+        for index, category in enumerate(categoryList):
+            info('crawling category(%s of %s): %s @ %s' %((index+1), len(categoryList), category, category.url))
+            self.crawlProducts(category)
+        
+    def crawlProducts(self, category_info):
+        url = category_info.url
+        category_idx, page_idx = 0, 0
+        success_count, fail_count, total_count = 0, 0, 0
+        while True:
+            num_idx = 0
+            products_page_content = self.fetchProductListPageContent(url)
+            page_idx += 1
+            parsed_products = self.parser.parseProductsByCategory(products_page_content, category_info)
+            for productInfo in parsed_products:
+                num_idx += 1
+                category_idx += 1
+                try:
+                    if self.parser.needProductDetails():
+                        product_page_content = self.fetcher.fetchProductPageContent(productInfo['product_url'])
+                        self.parser.parseProductDetails(product_page_content, productInfo)
+                    if self.parser.needRankInfo():
+                        productInfo['page_idx'] = str(page_idx)
+                        productInfo['num_idx'] = str(num_idx)
+                        productInfo['category_index'] = str(category_idx)
+                    print_csvline(self.fp, productInfo.to_list())
+                    success_count += 1
+                except:
+                    fail_count += 1
+                finally:
+                    total_count += 1
+                    if(total_count % 100 == 0):
+                        info("Below current category, crawled products: %s (%s success, %s failed)" %(total_count, success_count, fail_count))
+            url = self.parser.parseNextPageUrl(products_page_content) if len(parsed_products) > 4 else None
+            if not url:
+                break
+            if page_idx >= self.merchant.maxPagesPerCategory: #等于10表示第十页抓完
+                break
+        info("category: %s, totally crawled products: %s (%s success, %s failed)" %(category_info, total_count, success_count, fail_count))
+
     def fetchCategoryPageContent(self, furl):
         fetchTask = FetchTask(url=furl,
                               encoding='utf-8', 
                               headers=self.merchant.headers)
-        status, content = self.getHtmlDoc(fetchTask, 5)
+        info("fetching category page @ %s" %furl)
+        status, content = self.fetcher.getHtmlDoc(fetchTask, 5)
         return content
     
     def fetchProductListPageContent(self, furl):
         fetchTask = FetchTask(url=furl,
                               encoding='utf-8',
                               headers={})
-        status, content = self.getHtmlDoc(fetchTask, 3)
+        info("fetching productList page @ %s" %furl)
+        status, content = self.fetcher.getHtmlDoc(fetchTask, 3)
         return content
 
     def fetchProductPageContent(self, furl):
         fetchTask = FetchTask(url=furl,
                               encoding='utf-8',
                               headers={})
-        status, content = self.getHtmlDoc(fetchTask, 3)
+        info("fetching product page @ %s" %furl)
+        status, content = self.fetcher.getHtmlDoc(fetchTask, 3)
         return content
     
     def fetchPageWithoutHeaders(self, furl):
         fetchTask = FetchTask(url=furl,
                               encoding='utf-8',
                               headers={})
-        status, content = self.getHtmlDoc(fetchTask, 3)
+        info("fetching page without headers @ %s" %furl)
+        status, content = self.fetcher.getHtmlDoc(fetchTask, 3)
         return content
     
     def fetchSocialLikes(self, url):
         furl = 'http://graph.facebook.com/' + url
+        info('fetching facebook likes of product @ %s', url)
         content = self.fetchPageWithoutHeaders(furl)
         datas = json.loads(content)
         if datas.has_key('shares'):
             return datas['shares']
-
-class Crawler:
-    def __init__(self, merchant, parser):
-        self.merchant = merchant
-        self.fetcher = Fetcher(merchant)
-        self.parser = parser
-        self.filename = merchant.merchantName + '_' + datetime.date.today().strftime("%m-%d-%Y") + '_productInfo.inprogress'
-        self.fw = open(self.filename, 'w')
-
-    def crawl(self):
-        categoryList = self.getInitialCategories()
-        self.crawlCategoryList(categoryList)
-        self.fw.close()
-
-    def getInitialCategories(self):
-        homepage_content =self.fetcher.fetchCategoryPageContent(self.merchant.crawlEntryUrl)
-        categoryList = self.parser.parseCategories(homepage_content)
-        return categoryList
-    
-    def crawlCategoryList(self, categoryList):
-        for category in categoryList:
-            self.crawlProducts(category)
-        
-    def crawlProducts(self, category_info):
-        url = category_info.url
-        category_idx, page_idx = 0, 0
-        while True:
-            num_idx = 0
-            products_page_content = self.fetcher.fetchProductListPageContent(url)
-            page_idx += 1
-            parsed_products = self.parser.parseProductsByCategory(products_page_content, category_info)
-            for productInfo in parsed_products:
-                num_idx += 1
-                category_idx += 1
-                if self.parser.needProductDetails():
-                    product_page_content = self.fetcher.fetchProductPageContent(productInfo['product_url'])
-                    self.parser.parseProductDetails(product_page_content, productInfo)
-                if self.parser.needRankInfo():
-                    productInfo['page_idx'] = str(page_idx)
-                    productInfo['num_idx'] = str(num_idx)
-                    productInfo['category_index'] = str(category_idx)
-                print_csvline(self.fw, productInfo.to_list())
-            url = self.parser.parseNextPageUrl(products_page_content)
-            if not url:
-                break
 
 class Parser:
     def __init__(self, merchant):
@@ -227,29 +329,60 @@ class Parser:
     
     def newProduct(self):
         return ProductInfo(self.merchant)
-    
 
-def init_merchant(merchantName):
-    configFile = './config/' + merchantName + '.cfg'
-    config = read_config(configFile)
+LEVELS = {
+          'DEBUG': logging.DEBUG,
+          'INFO': logging.INFO,
+          'WARNING': logging.WARNING,
+          'ERROR': logging.ERROR
+          }    
+
+def logging_params(merchantName, config):
+    if not os.path.exists('logs'):
+        os.mkdir('logs')
+    logFilename = './logs/' + merchantName + '.log'
+    if config.get2('logLevel'):
+        logLevel = LEVELS[config.get2('logLevel')]
+    else:
+        logLevel = logging.INFO
+    return logFilename, logLevel
+    
+def config_log(merchantName, config):
+    logFilename, logLevel = logging_params(merchantName, config)
+    logging.basicConfig(filename=logFilename,
+                        level=logLevel,
+                        format="%(asctime)s [%(levelname)s] %(message)s",
+                        datefmt="%a %Y/%m/%d %H:%M:%S",
+                        filemode="a")
+
+def init_merchant(merchantName, config):
     siteUrl = config.get2("url")
+    #避免config未填写url配置信息报错
     if not siteUrl:
         siteUrl = 'www.' + merchantName + '.com'
     m = Merchant(merchantName, siteUrl)
     m.crawlEntryUrl = config.get2("crawlEntryUrl")
     m.productFields = config.getList("productFields")
+    m.sleepSeconds = tuple(item for item in config.getList('sleepSeconds'))
     headersFile = config.get2("headersFile")
     if headersFile:
         m.headers = load_http_headers('./config/' + headersFile)
+    if config.get2('maxPagesPerCategory'):
+        m.maxPagesPerCategory = int(config.get2('maxPagesPerCategory'))
+    m.useWebClient = config.getBoolean('useWebClient')
     m.config = config
+    if config.getBoolean('isSavePages'):
+        m.initStoreDir()
     return m
 
-def crawl(merchantName):
-    merchant = init_merchant(merchantName)
+def crawl(merchantName, config):
+    merchant = init_merchant(merchantName, config)
+    info("init parser...")
     moduleName, parserClassName = merchant.config.get2('parser').split(".")
     module = __import__(moduleName)
     parserClass = getattr(module, parserClassName)
     parser = parserClass(merchant)
+    info("init crawler...")
     crawler = Crawler(merchant, parser)
     parser.crawler = crawler  #方便调用一些方法
     crawler.crawl()
@@ -259,23 +392,25 @@ if __name__ == '__main__':
         print >> sys.stderr, 'Usage: python crawler2.py <merchantName>'
         sys.exit(1)
     merchantName = sys.argv[1]
-    #1. 一次完整的运行总是会运行crawl()方法, 并将锁解除
-    #2. 文件目录下无锁时, 生成新锁, 最后release()
-    #3. 文件目录下有锁时, 使用就锁, 最后remove()
+    #应该将设置log提前
+    configFile = './config/' + merchantName + '.cfg'
+    config = read_config(configFile)
+    config_log(merchantName, config)
     try:
-        isLockCrawlMerchant = True
-        if isLockCrawlMerchant:
-            lock = lockfile.FileLock("crawl-" + merchantName)
-            info("Try to acquire a lock of merchant <%s>..." %merchantName)
-            try:
-                lock.acquire(10)
-                info("Well done, we generate a new lock.")
-            except Exception, e:
+        checkLock = False #生产环境应该明确为True
+        lock = lockfile.FileLock("crawl-" + merchantName)
+        info("Try to acquire a lock of merchant <%s>..." %merchantName)
+        try:
+            lock.acquire(10)
+            info("Well done, we generate a new lock.")
+        except Exception, e:#Timeout waiting to acquire lock for
                 warning("A lock with same name exists, we just use it as new lock or stop our work?")
-                debug(e)  #Timeout waiting to acquire lock for
-        else:
-            info("Maybe we are testing, without using lock...")
-        crawl(merchantName)
+                if checkLock:
+                    warning("Yes, we just stop the work.")
+                    sys.exit(1)
+                else:
+                    info("No, we just use the old one and go on running.")
+        crawl(merchantName, config)
     except Exception, e:
         print e
     finally:
@@ -284,6 +419,7 @@ if __name__ == '__main__':
                 lock.release() #lockfile.NotMyLock: crawl-xxx is locked, but not by me
                 info('New lock release!')
         except lockfile.NotMyLock, e:
-            shutil.rmtree('crawl-' + merchantName + '.lock', True)
-            info('After work, we remove the old lock with same name!')
-
+            if not checkLock: #除非明确要求, 否则我们不该随便删除已经存在的锁
+                shutil.rmtree('crawl-' + merchantName + '.lock', True)
+                info('After work, we remove the old lock with same name!')
+        info("\n\n")
